requirements.txt
# Capsule - Local Knowledge OS
# requirements.txt - Version 1.1.0

fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
pgvector==0.3.5
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.12
mcp[cli]>=1.0.0
